+++ web/angular.json
@@ -39,7 +39,7 @@
               "node_modules/jquery/dist/jquery.slim.js",
               "node_modules/bootstrap/dist/js/bootstrap.js",
               "node_modules/quill/dist/quill.js",
-              "src/assets/scripts/navbar.js"
+              "src/assets/scripts/rwd-navbar.js"
             ]
           },
           "configurations": {
