+++ web/src/app/app.module.ts
@@ -1,92 +1,66 @@
-import {NgModule} from '@angular/core';
-import {BrowserModule} from '@angular/platform-browser';
-import {AppRoutingModule} from "./app-routing.module";
-import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {WelcomeComponent} from "./components/welcome/welcome.component";
+import {DoBootstrap, NgModule} from '@angular/core';
+import {HashLocationStrategy, LocationStrategy} from "@angular/common";
+import {TosidropNavbarComponent} from "./tosidrop/components/navbar/tosidrop-navbar.component";
+import {RewardzNavbarComponent} from "./rewardz/components/navbar/rewardz-navbar.component";
+import {AppTosidropRoutingModule} from "./tosidrop/app-tosidrop-routing.module";
+import {AppRewardzRoutingModule} from "./rewardz/app-rewardz-routing.module";
+import {AppTosidropModule} from "./tosidrop/app-tosidrop.module";
+import {AppRewardzModule} from "./rewardz/app-rewardz.module";
+import {Route, RouterModule, Routes} from "@angular/router";
 import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
 import {HttpClientModule} from "@angular/common/http";
-import {HashLocationStrategy, LocationStrategy} from "@angular/common";
-import {NavbarComponent} from "./components/navbar/navbar.component";
-import {TableModule} from "primeng/table";
+import {FormsModule, ReactiveFormsModule} from "@angular/forms";
 import {ModalModule} from "ngx-bootstrap/modal";
 import {TooltipModule} from "ngx-bootstrap/tooltip";
-import {RouterModule} from "@angular/router";
-import {NotifierModule, NotifierOptions} from "angular-notifier";
 import {OverlayPanelModule} from "primeng/overlaypanel";
-import {DragDropModule} from "@angular/cdk/drag-drop";
+import {CheckboxModule} from "primeng/checkbox";
+import {TableModule} from "primeng/table";
 import {DropdownModule} from "primeng/dropdown";
 import {MenubarModule} from "primeng/menubar";
-import {CheckboxModule} from "primeng/checkbox";
 import {DataViewModule} from "primeng/dataview";
-import {FooterComponent} from "./components/footer/footer.component";
-import {RestService} from "./services/rest.service";
-import {ButtonModule} from "primeng/button";
-import {RippleModule} from "primeng/ripple";
 import {ImageModule} from "primeng/image";
-import {InfoComponent} from "./components/welcome/info/info.component";
-import {DashboardComponent} from "./components/dashboard/dashboard.component";
-import {RewardsComponent} from "./components/dashboard/rewards/rewards.component";
-import {FeedbackComponent} from "./components/dashboard/feedback/feedback.component";
-import {HistoryComponent} from "./components/dashboard/history/history.component";
-import {WalletObserverService} from "./services/wallet-observer.service";
+import {ButtonModule} from "primeng/button";
+import {RestService} from "./common/services/rest.service";
+import {WalletObserverService} from "./common/services/wallet-observer.service";
+import {BrowserModule} from "@angular/platform-browser";
+import {WalletService} from "./common/services/wallet.service";
+const routes: Route[] = [
 
-const customNotifierOptions: NotifierOptions = {
-    position: {
-        horizontal: {
-            position: 'right',
-            distance: 12
-        },
-        vertical: {
-            position: 'bottom',
-            distance: 12,
-            gap: 10
-        }
-    },
-    theme: 'material',
-    behaviour: {
-        autoHide: 5000,
-        onClick: false,
-        onMouseover: 'pauseAutoHide',
-        showDismissButton: false,
-        stacking: 5
-    },
-    animations: {
-        enabled: true,
-        show: {
-            preset: 'slide',
-            speed: 300,
-            easing: 'ease'
-        },
-
-        hide: {
-            preset: 'fade',
-            speed: 300,
-            easing: 'ease',
-            offset: 50
-        },
-
-        shift: {
-            speed: 300,
-            easing: 'ease' // All standard CSS easing methods work
-        },
-        overlap: 150
-    }
-};
+];
 
 @NgModule({
-    declarations: [WelcomeComponent, NavbarComponent, FooterComponent, InfoComponent, DashboardComponent, RewardsComponent,
-    FeedbackComponent, HistoryComponent],
-    imports: [BrowserModule, FormsModule, ReactiveFormsModule, AppRoutingModule, RouterModule,
-        ModalModule.forRoot(), TooltipModule.forRoot(), OverlayPanelModule, DragDropModule,
-        BrowserAnimationsModule, HttpClientModule, CheckboxModule, TableModule,
-        DropdownModule, MenubarModule, DataViewModule, ButtonModule, RippleModule, ImageModule,
-        NotifierModule.withConfig(customNotifierOptions),
-    ],
-    bootstrap: [NavbarComponent],
-    providers: [RestService, WalletObserverService, {
+    imports: [BrowserModule, BrowserAnimationsModule, HttpClientModule,   RouterModule.forRoot(routes), RouterModule,
+        FormsModule, ReactiveFormsModule,
+        ModalModule.forRoot(), TooltipModule.forRoot(), OverlayPanelModule,
+        CheckboxModule, TableModule,
+        DropdownModule, MenubarModule, DataViewModule, ImageModule,
+        ButtonModule],
+    entryComponents: [TosidropNavbarComponent, RewardzNavbarComponent],
+    providers: [RestService, WalletObserverService, WalletService, {
         provide: LocationStrategy,
         useClass: HashLocationStrategy
-    }]
+    }],
 })
-export class AppModule {
+
+
+export class AppModule implements DoBootstrap {
+
+    ngDoBootstrap(app) {
+        document.querySelector('#pre-load').remove();
+        if (!window.location.pathname.toLowerCase().includes("tosidrop")) {
+            console.log("tosidrop");
+            const componentElement = document.createElement("tosidrop");
+            document.body.appendChild(componentElement);
+            import("./tosidrop/app-tosidrop.module");
+            import("./tosidrop/app-tosidrop-routing.module");
+            app.bootstrap(TosidropNavbarComponent);
+        } else {
+            console.log("rewardz");
+            const componentElement = document.createElement("rewardz");
+            document.body.appendChild(componentElement);
+            import("./rewardz/app-rewardz.module");
+            import("./rewardz/app-rewardz-routing.module");
+            app.bootstrap(RewardzNavbarComponent);
+        }
+    }
 }
