+++ web/src/app/common/data/available-tokens.ts
