+++ web/src/app/common/data/script.ts
