+++ web/src/app/common/services/wallet-observer.service.ts
