+++ web/src/app/rewardz/app-rewardz-routing.module.ts
@@ -1,23 +1,22 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes, PreloadAllModules} from '@angular/router';
-import {WelcomeComponent} from "./components/welcome/welcome.component";
-import {DashboardComponent} from "./components/dashboard/dashboard.component";
 import {RewardsComponent} from "./components/dashboard/rewards/rewards.component";
 import {HistoryComponent} from "./components/dashboard/history/history.component";
 import {FeedbackComponent} from "./components/dashboard/feedback/feedback.component";
+import {DashboardComponent} from "./components/dashboard/dashboard.component";
+import {RewardzWelcomeComponent} from "./components/welcome/rewardz-welcome.component";
 
 const routes: Routes = [
     {
         path: '',
-        redirectTo: 'welcome',
-        pathMatch: 'full'
+        redirectTo: '/welcome',
     },
     {
-        path: 'welcome',
-        component: WelcomeComponent
+        path: '/welcome',
+        component: RewardzWelcomeComponent,
     },
     {
-        path: 'dashboard',
+        path: '/dashboard',
         component: DashboardComponent,
         children: [
             {path: '', redirectTo: 'rewards', pathMatch: 'full'},
@@ -31,11 +30,10 @@ const routes: Routes = [
 @NgModule({
     imports: [RouterModule.forRoot(routes, {
         useHash: true,
-        preloadingStrategy: PreloadAllModules,
         scrollPositionRestoration: "enabled"
     })],
     exports: [RouterModule]
 })
 
-export class AppRoutingModule {
+export class AppRewardzRoutingModule {
 }
