+++ web/src/app/rewardz/components/footer/footer.component.ts
@@ -2,7 +2,7 @@ import {Component, Injectable} from "@angular/core";
 
 @Component({
     selector: 'footer',
-    styleUrls: ['../../../styles/page-content.css'],
+    styleUrls: ['../../styles/page-content.css'],
     templateUrl: './footer.html',
 })
 @Injectable()
