+++ web/src/app/rewardz/components/navbar/rewardz-navbar.html
@@ -29,7 +29,7 @@
             <div class="modal-header">
                 <h4 class="modal-title ms-0 me-auto">Connect a wallet</h4>
                 <span class="ms-auto">
-                <button pButton pRipple class="p-button-rounded p-button-text p-button-plain" icon="pi pi-times"
+                <button pButton class="p-button-rounded p-button-text p-button-plain" icon="pi pi-times"
                         title="Cancel"
                         iconPos="left"
                         (click)="hideConnectModal()"></button>
@@ -37,28 +37,28 @@
             </div>
             <div class="modal-body text-center" style="padding: .1em !important;overflow-y: auto">
                 <div class="ms-2 me-2 row">
-                    <a pButton pRipple class="p-button-rounded p-button-outlined no-ul inverted"
+                    <a pButton class="p-button-rounded p-button-outlined no-ul inverted"
                        [title]="ccvaultAvailable() ? 'Connect to CCVault':'CCVault wallet not installed'"
                        (click)="connectCCVault()">
                         <p class="ms-1">CCVault</p>
                         <p-image src="../../assets/icons/ccvault.png" alt="CCVault" width="25" height="25">
                         </p-image>
                     </a>
-                    <a pButton pRipple class="p-button-rounded p-button-outlined no-ul inverted mt-3"
+                    <a pButton class="p-button-rounded p-button-outlined no-ul inverted mt-3"
                        (click)="connectNami()"
                        [title]="namiAvailable() ? 'Connect to Nami':'Nami wallet not installed'">
                         <p class="ms-1">Nami</p>
                         <p-image src="../../assets/icons/nami.png" alt="Nami" width="25" height="25">
                         </p-image>
                     </a>
-                    <a pButton pRipple class="p-button-rounded p-button-outlined no-ul inverted mt-3"
+                    <a pButton class="p-button-rounded p-button-outlined no-ul inverted mt-3"
                        (click)="connectGero()"
                        [title]="geroAvailable() ? 'Connect to Gero':'Gero wallet not installed'">
                         <p class="ms-1">Gero</p>
                         <p-image src="../../assets/icons/gero.png" alt="Gero" width="25" height="25">
                         </p-image>
                     </a>
-                    <a pButton pRipple class="p-button-rounded p-button-outlined no-ul inverted mt-3"
+                    <a pButton class="p-button-rounded p-button-outlined no-ul inverted mt-3"
                        (click)="connectFlint()"
                        [title]="flintAvailable() ? 'Connect to Flint':'Flint wallet not installed'">
                         <p class="ms-1">Flint</p>
