+++ web/src/app/rewardz/components/welcome/info/info.component.ts
@@ -2,7 +2,7 @@ import {Component, Injectable} from '@angular/core';
 
 @Component({
     selector: 'info',
-    styleUrls: ['../../../../styles/page-content.css'],
+    styleUrls: ['../../../styles/page-content.css'],
     templateUrl: './info.html'
 })
 @Injectable()
