+++ web/src/app/rewardz/components/welcome/rewardz-welcome.component.ts
@@ -2,13 +2,13 @@ import {Component} from '@angular/core';
 
 @Component({
     selector: 'welcome',
-    styleUrls: ['../../../styles/page-content.css'],
-    templateUrl: './welcome.html'
+    styleUrls: ['../../styles/page-content.css'],
+    templateUrl: './rewardz-welcome.html'
 })
 /**
  * Main Dashboard for the queries
  */
-export class WelcomeComponent {
+export class RewardzWelcomeComponent {
 
     constructor() {
     }
