+++ web/src/app/rewardz/styles/overrides.css
@@ -12,7 +12,7 @@ body {
     overflow-x: hidden !important;
     overscroll-behavior: none !important;
     padding-top: 0px !important;
-    background: url('../assets/rwd-background.png') no-repeat center center fixed;
+    background: url('../../../assets/rwd-background.png') no-repeat center center fixed;
     -webkit-background-size: cover;
     -moz-background-size: cover;
     -o-background-size: cover;
