+++ web/src/app/rewardz/styles/responsive.css
@@ -1,0 +1,33 @@
+/* This CSS files overrides and supplements Bootstrap's responsive design */
+@media only screen and (min-width: 960px) {
+    /* styles for browsers larger than 960px !important; */
+}
+
+@media only screen and (min-width: 1200px) {
+    /* styles for browsers larger than 1200px !important; */
+}
+
+@media only screen and (min-width: 1440px) {
+    /* styles for browsers larger than 1440px !important; */
+}
+
+@media only screen and (min-width: 1900px) {
+    /* for sumo sized (mac) screens */
+}
+
+@media only screen and (max-device-width: 480px) {
+    /* styles for mobile browsers smaller than 480px !important; (iPhone) */
+}
+
+@media only screen and (device-width: 768px) {
+    /* default iPad screens */
+}
+
+/* different techniques for iPad screening */
+@media only screen and (min-device-width: 481px) and (max-device-width: 1024px) and (orientation: portrait) {
+    /* For portrait layouts only */
+}
+
+@media only screen and (min-device-width: 481px) and (max-device-width: 1024px) and (orientation: landscape) {
+    /* For landscape layouts only */
+}
