+++ web/src/app/tosidrop/components/navbar/tosidrop-navbar.html
@@ -1,0 +1,17 @@
+<nav class="navbar navbar-expand-lg fixed-top navbar-dark bg-dark bg-clear">
+    <div class="container-fluid">
+        <ul  class="navbar-nav ms-2 mb-2">
+            <li class="nav-item"><a class="nav-link" href="#"><i class="fa-solid fa-home me-1"></i> TOSIDROP</a></li>
+        </ul>
+        <ul  class="navbar-nav ms-auto me-2 mb-2 mb-lg-0">
+            <li class="nav-item"><a class="nav-link" ><i class="fa-solid fa-wallet me-1"></i>Connect</a>
+            </li>
+            <li class="nav-item"><a class="nav-link" href="#contact"><i class="fa-solid fa-question-circle me-1"></i>HELP</a>
+            </li>
+        </ul>
+    </div>
+</nav>
+<notifier-container></notifier-container>
+<div class="container-fluid" style="padding-top: 75px">
+    <router-outlet></router-outlet>
+</div>
