+++ web/src/app/tosidrop/components/welcome/tosidrop-welcome.component.ts
@@ -1,0 +1,15 @@
+import {Component} from '@angular/core';
+
+@Component({
+    selector: 'welcome',
+    styleUrls: ['../../styles/page-content.css'],
+    templateUrl: './tosidrop-welcome.html'
+})
+/**
+ * Main Dashboard for the queries
+ */
+export class TosidropWelcomeComponent {
+
+    constructor() {
+    }
+}
