+++ web/src/app/tosidrop/components/welcome/tosidrop-welcome.html
@@ -1,0 +1,13 @@
+<header class="intro">
+    <div class="intro-body">
+        <div class="container">
+            <div class="row">
+                <h1 class="brand-heading">Welcome to Tosidrop</h1>
+                <p class="intro-text">A Tosidrop Project<br>
+                    <a href="#info" class="btn btn-circle page-scroll">
+                        <i class="fa-solid fa-angle-double-down animated"></i>
+                    </a>
+            </div>
+        </div>
+    </div>
+</header>
