+++ web/src/index.html
@@ -7,19 +7,13 @@
     <meta charset="UTF-8">
     <meta name="viewport"
           content="width=device-width, initial-scale=1, minimum-scale=1, maximum-scale=1, user-scalable=no">
-    <!-- Add hostname to title -->
-    <script>
-        document.title = "Rewardz";
-    </script>
     <!-- Display the application -->
 </head>
 <body>
-<navbar>
     <div id="pre-load">
         <div class="messaging">
-        <!--    <img class="blinking20" src="./assets/loading.png" height="300" width="600"/>-->
+            <!--    <img class="blinking20" src="./assets/loading.png" height="300" width="600"/>-->
         </div>
     </div>
-</navbar>
 </body>
 </html>
